requirements.txt
openai-agents>=0.13.1
fastapi
uvicorn[standard]
python-dotenv
